+++ examples/base_cc_gwlb_zpa/main.tf
@@ -195,6 +195,9 @@ module "gwlb" {
   healthy_threshold        = var.healthy_threshold
   unhealthy_threshold      = var.unhealthy_threshold
   cross_zone_lb_enabled    = var.cross_zone_lb_enabled
+  deregistration_delay     = var.deregistration_delay
+  flow_stickiness          = var.flow_stickiness
+  rebalance_enabled        = var.rebalance_enabled
 }