+++ modules/terraform-zscc-gwlb-aws/main.tf
@@ -2,11 +2,12 @@
 # Configure target group
 ################################################################################
 resource "aws_lb_target_group" "gwlb_target_group" {
-  name        = var.target_group_name
-  port        = 6081
-  protocol    = "GENEVE"
-  vpc_id      = var.vpc_id
-  target_type = var.asg_enabled == true ? "instance" : "ip"
+  name                 = var.target_group_name
+  port                 = 6081
+  protocol             = "GENEVE"
+  vpc_id               = var.vpc_id
+  target_type          = var.asg_enabled == true ? "instance" : "ip"
+  deregistration_delay = var.deregistration_delay
 
   health_check {
     port                = var.http_probe_port
@@ -15,6 +16,18 @@ resource "aws_lb_target_group" "gwlb_target_group" {
     interval            = var.health_check_interval
     healthy_threshold   = var.healthy_threshold
     unhealthy_threshold = var.unhealthy_threshold
+  }
+
+  target_failover {
+    on_deregistration = var.rebalance_enabled == true ? "rebalance" : "no_rebalance"
+    on_unhealthy      = var.rebalance_enabled == true ? "rebalance" : "no_rebalance"
+  }
+
+  # type attribute only applies if enabled = true and only options are "source_ip_dest_ip" (2-tuple) or "source_ip_dest_ip_proto" (3-tuple).
+  # enabled = false implies 5-tuple. AWS gives type a default value of "source_ip_dest_ip_proto" even if enabled is set to false
+  stickiness {
+    enabled = var.flow_stickiness == "5-tuple" ? false : true
+    type    = var.flow_stickiness == "2-tuple" ? "source_ip_dest_ip" : "source_ip_dest_ip_proto"
   }
 }